Move secrets config sources to correct directories. Refactor secrets config sources to use Service Registry (from new `helidon-integrations-oci` module)
