+++ integrations/oci/pom.xml
@@ -35,8 +35,8 @@
     <modules>
         <module>oci</module>
         <module>metrics</module>
-        <module>oci-secrets-config-source</module>
-        <module>oci-secrets-mp-config-source</module>
+        <module>secrets-config-source</module>
+        <module>secrets-mp-config-source</module>
         <module>sdk</module>
         <module>tls-certificates</module>
     </modules>
